refactor(app): extract session turn changes
